requirements.txt
uvicorn
fastapi
starlette
sqlalchemy[asyncio]
databases
psycopg2
asyncpg
python-dotenv
pydantic-settings
alembic
google-auth
google-auth-oauthlib
google-auth-httplib2
pyjwt
python-jose[cryptography]
httpx
apscheduler
google-adk